main's ENG-5172 refactor replaced normalizeOrg() string with
normalizeOrgTokens() []string and swapped containment similarity for
JaccardTokenSets plus a TokenSetContained guard. The dotted-acronym
collapse moves into normalizeOrgTokens, and the expected verdicts are
recomputed under Jaccard: the two "longer registrant" cases now land on
unverifiable rather than match, since a single-token pivot contained in a
four-token registrant scores 0.25. That is still the fix's point — without
the collapse the leftover single letters are private tokens on one side,
which defeats the containment guard and produces an affirmative mismatch.

Also bounds dottedAcronymPattern per PR review (Codex, Gemini):

  (?i)\b(?:[a-z]\.){2,}  ->  (?i)\b(?:[a-z]\.)+[a-z]\b

Counting letter-period pairs skipped a two-letter suffix written without
its trailing period ("Globex S.A" carries one period, so {2,} never
matched and the suffix was never stripped) and, lacking a right-hand
boundary, ran on into a following word ("U.S.Army" -> "usarmy"). Closing
the run on a letter at a word boundary fixes both while leaving the
blanket-strip tripwires -- Acme.Corp, acme.com, J.Crew, John Q. Public --
byte-identical.

+++ pkg/plugins/plugin.go
@@ -74,10 +74,10 @@ type Input struct {
 
 // Confidence is a single piece of scored, explained evidence supporting a
 // finding. Entries are additive: a finding's total confidence is the sum of its
-// entry scores, capped at 1.0 (see TotalConfidence).
+// entry scores, capped at 100 (see TotalConfidence).
 type Confidence struct {
 	// Score is this evidence's contribution to the total confidence.
-	Score float64 `json:"score"`
+	Score int `json:"score"`
 
 	// Justification explains, in human-readable terms, why this evidence
 	// supports the finding.
@@ -122,7 +122,7 @@ type findingFields Finding
 func (f Finding) MarshalJSON() ([]byte, error) {
 	return json.Marshal(struct {
 		findingFields
-		TotalConfidence float64
+		TotalConfidence int
 		NeedsReview     bool
 	}{
 		findingFields:   findingFields(f),
